vector_db/requirements.txt
faiss-cpu
langchain
fastapi
uvicorn
langchain_community
langchain_google_genai
python-dotenv
